+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file.
 - Added an ability to specify relative/absolute theme path in config file
 - Added `keep_state_on_song_change` and equivalent flag to cli
 - Added `ignore_leading_the` when sorting entries in browsers
+- More information about the system to debuginfo
 
 ### Changed
 
@@ -54,6 +55,8 @@ remote commands inside scripts triggered by rmpc
 `keep_state_on_song_change` to disable this 
 - Default theme now includes lyrics pane above the album art on queue tab
 - Browsers now properly use case insensitive sorting
+- Refactored and improved image backend detection
+- `JumpToCurrent` now jumps to last playing song in stopped state
 
 ### Fixed
 
@@ -66,6 +69,9 @@ scrolled down
 - Fix improper handling of remote theme change
 - Tilde not being expanded for yt-dlp on non-linux platforms
 - Remove mention of `tab_bar.enabled` from docs
+- Konsole terminal now does not autodetect to Kitty image protocol, it instead uses ueberzugpp if
+available and Block if not
+- Fix Iterm2 image protocol sometimes rendering too late
 
 ## [0.9.0] - 2025-06-23